@@ -13,11 +13,12 @@
     }
   ],
   "require": {
-    "php": ">=7.0",
+    "php": ">=7.1",
+    "ext-pdo": "*",
+    "ext-json": "*",
     "twig/twig":"~2.0",
     "phpunit/phpunit": "^6.0",
     "predis/predis":"1.1.*@dev",
-    "codacy/coverage": "dev-master",
     "auth0/auth0-php": "^5.0"
   },
   "suggest": {
@@ -27,5 +28,8 @@
     "psr-0": {
       "Staple": "library/"
     }
+  },
+  "require-dev": {
+    "codacy/coverage": "dev-master"
   }
 }
