@@ -24,6 +24,7 @@
 namespace Staple\Query;
 
 use Exception;
+use Staple\Exception\ConfigurationException;
 use Staple\Exception\QueryException;
 
 class Condition
@@ -40,6 +41,10 @@ class Condition
 	const BETWEEN = "BETWEEN";
 	const LIKE = "LIKE";
 	const NOTLIKE = "NOT LIKE";
+	const PARAMETERIZED_QUERY = true;
+	const NON_PARAMETERIZED_QUERY = false;
+	const SQL_AND = 'AND';
+	const SQL_OR = 'OR';
 
 	/**
 	 * The column for the where
@@ -57,6 +62,26 @@ class Condition
 	 */
 	protected $value;
 	/**
+	 * An array of values used for IN and BETWEEN operators
+	 * @var array
+	 */
+	protected $valueArray;
+	/**
+	 * The name of the parameter for a parameterized query.
+	 * @var string
+	 */
+	protected $paramName;
+	/**
+	 * An array of parameter names used for the IN and BETWEEN operators
+	 * @var string[]
+	 */
+	protected $paramArray;
+	/**
+	 * Flag to
+	 * @var bool
+	 */
+	protected $parameterized = true;
+	/**
 	 * An override statement that represents the WHERE clause.
 	 * @var string
 	 */
@@ -66,6 +91,12 @@ class Condition
 	 * @var bool
 	 */
 	protected $columnJoin = false;
+
+	/**
+	 * SQL Conditional Conjunction (AND/OR)
+	 * @var string
+	 */
+	protected $conjunction;
 	/**
 	 * Reference to the connection object in use for this clause.
 	 * @var Connection
@@ -111,7 +142,12 @@ class Condition
 		$this->statement = (string)$where;
 		return $this;
 	}
-	
+
+	/**
+	 * @param Connection $connection
+	 * @return string
+	 * @throws QueryException
+	 */
 	public function build(Connection $connection)
 	{
 		if(isset($this->statement))
@@ -123,15 +159,33 @@ class Condition
 			if(strtoupper($this->operator) == self::IN)
 			{
 				$value = "(";
+
 				if(is_array($this->value))
 				{
-					foreach($this->value as $aValue)
+					if($this->isParameterized())
 					{
-						if(strlen($value) > 1)
+						$count = 0;
+						foreach($this->value as $aValue)
 						{
-							$value .= ",";
+							$count++;
+							$value .=
+								$this->columnJoin ?
+									$aValue :
+									(strlen($this->paramName) > 0) ?
+										':' . $this->getParamName() . '_in_' . $count  :
+										'?';
 						}
-						$value .= $this->columnJoin ? $aValue : Query::convertTypes($aValue,$connection);
+					}
+					else
+					{
+						foreach($this->value as $aValue)
+						{
+							if(strlen($value) > 1)
+							{
+								$value .= ", ";
+							}
+							$value .= $this->columnJoin ? $aValue : Query::convertTypes($aValue, $connection);
+						}
 					}
 				}
 				elseif($this->value instanceof Select)
@@ -140,17 +194,62 @@ class Condition
 				}
 				else
 				{
-					$value = $this->columnJoin ? $this->value : Query::convertTypes($this->value,$connection);
+					if($this->isParameterized())
+					{
+						$value .= (strlen($this->paramName) > 0) ? ':' . $this->getParamName() : '?';
+					}
+					else
+					{
+						$value .= $this->columnJoin ? $this->value : Query::convertTypes($this->value, $connection);
+					}
 				}
 				$value .= ")";
 			}
-			elseif (strtoupper($this->operator) == self::BETWEEN)
+			elseif(strtoupper($this->operator) === self::BETWEEN)
+			{
+				if($this->isParameterized())
+				{
+					if(strlen($this->paramArray['start']) > 0 && strlen($this->paramArray['end']) > 0)
+					{
+						$value = ':' . $this->paramArray['start'] . ' AND :'. $this->paramArray['end'];
+					}
+					else
+					{
+						$value = '? AND ?';
+					}
+				}
+				else
+				{
+					$value = $this->getValue();
+				}
+			}
+			//Use the column value in th
+			elseif($this->columnJoin === true)
 			{
 				$value = $this->getValue();
 			}
-			else 
+			//Don't parameterize a NULL value
+			elseif((strtoupper($this->operator) === self::IS || strtoupper($this->operator) === self::IS_NOT) && is_null($this->value))
 			{
-				$value = $this->columnJoin ? $this->value : Query::convertTypes($this->value,$connection);
+				$value = Query::convertTypes($this->value, $connection);
+			}
+			else
+			{
+				if($this->isParameterized())
+				{
+					if(strlen($this->paramName) > 0)
+					{
+						$value = ':' . $this->getParamName();
+					}
+					else
+					{
+						$value = '?';
+					}
+				}
+				else
+				{
+					$value = Query::convertTypes($this->value, $connection);
+				}
 			}
 			return $this->column.' '.$this->operator.' '.$value;
 		}
@@ -242,15 +341,72 @@ class Condition
 	 * @param bool $columnJoin
 	 * @return $this
 	 */
-	public function setColumnJoin($columnJoin)
+	public function setColumnJoin(bool $columnJoin)
 	{
-		$this->columnJoin = (bool)$columnJoin;
+		$this->columnJoin = $columnJoin;
+		return $this;
+	}
+
+	/**
+	 * @return string
+	 */
+	public function getParamName(): ?string
+	{
+		return $this->paramName;
+	}
+
+	/**
+	 * @param string $paramName
+	 * @return Condition
+	 */
+	public function setParamName(string $paramName): Condition
+	{
+		$this->paramName = $paramName;
+		return $this;
+	}
+
+	/**
+	 * @param string $paramName
+	 * @return Condition
+	 */
+	public function setStartParamName(string $paramName): Condition
+	{
+		$this->paramArray['start'] = $paramName;
+		return $this;
+	}
+
+	/**
+	 * @param string $paramName
+	 * @return Condition
+	 */
+	public function setEndParamName(string $paramName): Condition
+	{
+		$this->paramArray['end'] = $paramName;
+		return $this;
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function isParameterized(): bool
+	{
+		return $this->parameterized;
+	}
+
+	/**
+	 * @param bool $parameterized
+	 * @return Condition
+	 */
+	public function setParameterized(bool $parameterized): Condition
+	{
+		$this->parameterized = $parameterized;
 		return $this;
 	}
 
 	/**
 	 * Return the currently set connection or attempt to retrieve the default connection if non specified.
 	 * @return Connection
+	 * @throws ConfigurationException
 	 */
 	public function getConnection()
 	{
@@ -276,21 +432,37 @@ class Condition
 	/*-----------------------------------------------CONDITION ENCAPSULATORS-----------------------------------------------*/
 
 	/**
-	 * @param $column
+	 * @param string|Query $column
 	 * @param $operator
-	 * @param $value
-	 * @param bool $columnJoin
+	 * @param mixed $value
+	 * @param bool|null $columnJoin
+	 * @param string|null $paramName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return static
 	 */
-	public static function get($column, $operator, $value, $columnJoin = NULL)
+	public static function get($column, $operator, $value, bool $columnJoin = NULL, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator($operator)
-			->setValue($value);
+			->setValue($value)
+			->setParameterized($parameterized);
+
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Column Join
 		if(isset($columnJoin))
 			$obj->setColumnJoin($columnJoin);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
@@ -310,20 +482,35 @@ class Condition
 	 * @param string $column
 	 * @param mixed $value
 	 * @param bool $columnJoin
+	 * @param string $paramName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return Condition
 	 */
-	public static function equal($column, $value, $columnJoin = NULL)
+	public static function equal($column, $value, bool $columnJoin = NULL, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
-			->setValue($value);
+			->setValue($value)
+			->setParameterized($parameterized);
 		
 		//Check for NULLS
 		is_null($value) ? $obj->setOperator(self::IS) :	$obj->setOperator(self::EQUAL);
-		
+
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Column Join
 		if(isset($columnJoin))
 			$obj->setColumnJoin($columnJoin);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
@@ -332,38 +519,69 @@ class Condition
 	 * @param string $column
 	 * @param mixed $value
 	 * @param bool $columnJoin
+	 * @param string $paramName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return Condition
 	 */
-	public static function notEqual($column, $value, $columnJoin = NULL)
+	public static function notEqual($column, $value, bool $columnJoin = NULL, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
-			->setValue($value);
+			->setValue($value)
+			->setParameterized($parameterized);
 
 		//Check for NULLS
 		is_null($value) ? $obj->setOperator(self::IS_NOT) :	$obj->setOperator(self::NOTEQUAL);
 
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Column Join
 		if(isset($columnJoin))
 			$obj->setColumnJoin($columnJoin);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
 	/**
-	 * @param $column
-	 * @param $value
+	 * @param string $column
+	 * @param mixed $value
 	 * @param bool $columnJoin
+	 * @param string $paramName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return static
 	 */
-	public static function like($column, $value, $columnJoin = NULL)
+	public static function like($column, $value, $columnJoin = NULL, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator(self::LIKE)
-			->setValue($value);
+			->setValue($value)
+			->setParameterized($parameterized);
+
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Column Join
 		if(isset($columnJoin))
 			$obj->setColumnJoin($columnJoin);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
@@ -371,71 +589,119 @@ class Condition
 	 * @param $column
 	 * @param $value
 	 * @param bool $columnJoin
+	 * @param string|null $paramName
+	 * @param $conjunction
+	 * @param bool $parameterized
 	 * @return static
 	 */
-	public static function notLike($column, $value, $columnJoin = NULL)
+	public static function notLike($column, $value, bool $columnJoin = NULL, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator(self::NOTLIKE)
-			->setValue($value);
+			->setValue($value)
+			->setParameterized($parameterized);
+
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Columns Join
 		if(isset($columnJoin))
 			$obj->setColumnJoin($columnJoin);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
 	/**
-	 * @param $column
+	 * @param string|Query $column
+	 * @param string $conjunction
 	 * @return static
 	 */
-	public static function null($column)
+	public static function null($column, $conjunction = Condition::SQL_AND)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator(self::IS)
 			->setValue(NULL);
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
 	/**
 	 * @param $column
 	 * @param $values
-	 * @param bool $columnJoin
+	 * @param string|null $paramName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return static
 	 */
-	public static function in($column, $values, $columnJoin = NULL)
+	public static function in($column, $values, string $paramName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator(self::IN)
-			->setValue($values);
-		if(isset($columnJoin)) 
-			$obj->setColumnJoin($columnJoin);
+			->setValue($values)
+			->setParameterized($parameterized);
+
+		//Query Parameter Name
+		if(is_null($paramName))
+			$obj->setParamName(Query::sanitizeParamName($column));
+		else
+			$obj->setParamName(Query::sanitizeParamName($paramName));
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
+
 		return $obj;
 	}
 
 	/**
 	 * @param $column
-	 * @param $start
-	 * @param $end
-	 * @param IConnection $connection
+	 * @param mixed $start
+	 * @param mixed $end
+	 * @param string $startParamName
+	 * @param string $endParamName
+	 * @param string $conjunction
+	 * @param bool $parameterized
 	 * @return static
 	 * @throws QueryException
 	 */
-	public static function between($column, $start, $end, IConnection $connection = NULL)
+	public static function between($column, $start, $end, string $startParamName = null, string $endParamName = null, $conjunction = self::SQL_AND, bool $parameterized = true)
 	{
 		/** @var Condition $obj */
 		$obj = new static();
 		$obj->setColumn($column)
 			->setOperator(self::BETWEEN)
 			->setValue(Query::convertTypes($start)." AND ".Query::convertTypes($end))
-			->setColumnJoin(true);
-		
-		if(isset($connection))
-			$obj->setConnection($connection);
+			->setColumnJoin(true)
+			->setParameterized($parameterized);
+
+		//Start Parameter Name
+		if(is_null($startParamName))
+			$obj->setStartParamName(Query::sanitizeParamName($column.'_start'));
+		else
+			$obj->setStartParamName(Query::sanitizeParamName($startParamName));
+
+		//End Parameter Name
+		if(is_null($endParamName))
+			$obj->setEndParamName(Query::sanitizeParamName($column.'_end'));
+		else
+			$obj->setEndParamName(Query::sanitizeParamName($endParamName));
+
+		//Where Conjunction
+		$obj->conjunction = (strtoupper($conjunction) === self::SQL_OR) ? self::SQL_OR : self::SQL_AND;
 		
 		return $obj;
 	}
