@@ -23,7 +23,9 @@
 namespace Staple\Query;
 
 use Exception;
+use PDO;
 use Staple\Error;
+use Staple\Exception\ConfigurationException;
 use Staple\Exception\QueryException;
 use Staple\Traits\Factory;
 
@@ -83,18 +85,25 @@ class Insert
 	protected $columns;
 
 	/**
+	 * Set the flag for query parameterization
+	 * @var bool
+	 */
+	protected $parameterized = true;
+
+	/**
 	 * @param string $table
 	 * @param array $data
-	 * @param Connection $db
+	 * @param IConnection $db
 	 * @param string $priority
+	 * @param bool $parameterized
 	 * @throws QueryException
 	 */
-	public function __construct($table = NULL, $data = NULL, Connection $db = NULL, $priority = NULL)
+	public function __construct($table = null, $data = null, IConnection $db = null, $priority = null, bool $parameterized = null)
 	{
 		$this->data = new DataSet();
 		
 		//Process Database connection
-		if($db instanceof Connection)
+		if($db instanceof IConnection)
 		{
 			$this->setConnection($db);
 		}
@@ -104,7 +113,7 @@ class Insert
 				$db = Connection::get();
 				$this->setConnection($db);
 			}
-			catch (QueryException $e)
+			catch (ConfigurationException $e)
 			{
 				throw new QueryException('Unable to find a database connection.', Error::DB_ERROR, $e);
 			}
@@ -134,6 +143,12 @@ class Insert
 		{
 			$this->setPriority($priority);
 		}
+
+		//Set Priority
+		if(isset($parameterized))
+		{
+			$this->setParameterized($parameterized);
+		}
 	}
 	
 	/**
@@ -152,11 +167,17 @@ class Insert
 	}
 	
 	/**
-	 * 
 	 * @see Staple_Query::build()
+	 * @param bool $parameterized
+	 * @return string
+	 * @throws QueryException
+	 * @throws ConfigurationException
 	 */
-	function build()
+	function build(bool $parameterized = null)
 	{
+		if(isset($parameterized))
+			$this->setParameterized($parameterized);
+
 		//Statement Start
 		$stmt = "INSERT ";
 		
@@ -167,6 +188,10 @@ class Insert
 		}
 		if($this->ignore === TRUE)
 		{
+			if($this->connection->getDriver() === Connection::DRIVER_SQLITE)
+			{
+				$stmt .= 'OR ';
+			}
 			$stmt .= 'IGNORE ';
 		}
 		
@@ -185,7 +210,7 @@ class Insert
 		//Column List
 		if(isset($this->columns))
 		{
-			$stmt .= '('.implode(',',$this->getColumns()).') ';
+			$stmt .= '('.implode(', ', $this->getColumns()).') ';
 		}
 
 		//Data
@@ -211,7 +236,7 @@ class Insert
 				}
 				else
 				{
-					$stmt .= ',';
+					$stmt .= ', ';
 				}
 				$stmt .= " $ucol=VALUES($ucol)";
 			}
@@ -223,31 +248,134 @@ class Insert
 	/**
 	 * Executes the query.
 	 * @throws QueryException
-	 * @return Statement | bool
+	 * @throws ConfigurationException
+	 * @return IStatement | bool
 	 */
 	public function execute()
 	{
-		if($this->connection instanceof Connection)
+		//Make sure we have a connection object
+		if(!($this->connection instanceof Connection))
 		{
-			return $this->connection->query($this->build());
-		}
-		else
-		{
-			try 
+			try
 			{
 				$this->setConnection(Connection::get());
 			}
 			catch (Exception $e)
 			{
-				//@todo try for a default connection if no staple connection
 				throw new QueryException('No Database Connection', Error::DB_ERROR);
 			}
-			if($this->connection instanceof Connection)
-			{
-				return $this->connection->query($this->build());
-			}
 		}
-		return false;
+
+		if($this->parameterized !== true)
+		{
+			//Run raw query string.
+			return $this->connection->query($this->build());
+		}
+		else
+		{
+			$statement = $this->connection->prepare($this->build(true));
+			if($statement !== false && $statement !== null)
+			{
+				$this->bindParametersToStatement($statement);
+				if($statement->execute() === true)
+					return $statement;
+				else
+					throw new QueryException(json_encode($statement->errorInfo()));
+			}
+			else
+				throw new QueryException('Failed to prepare statement for execution.');
+		}
+	}
+
+	/**
+	 * @return array|string[]
+	 * @throws QueryException
+	 */
+	public function getParams()
+	{
+		if($this->data instanceof DataSet)
+		{
+			$dataSet = $this->data;
+
+			$data = array_filter($this->data->getData(), function($key) use($dataSet) {
+				return !$dataSet->isLiteral($key);
+			}, ARRAY_FILTER_USE_KEY );
+
+			return $data;
+		}
+		elseif($this->data instanceof Select)
+		{
+			return $this->data->getParams();
+		}
+		else
+		{
+			throw new QueryException('Insert Data is stored as a non-supported format.');
+		}
+	}
+
+	/**
+	 * Bind the query parameters to the supplied statement object.
+	 * @param IStatement $statement
+	 * @throws QueryException
+	 */
+	private function bindParametersToStatement(&$statement)
+	{
+		foreach($this->getParams() as $name=>$value)
+		{
+			$varType = gettype($value);
+
+			//Don't bind nulls because they are converted to IS NULL by the query builder.
+			if($varType === 'NULL')
+			{
+				continue;
+			}
+
+			//Bind other types.
+			switch($varType)
+			{
+				case "boolean":
+					$type = PDO::PARAM_BOOL;
+					break;
+				case "integer":
+					$type = PDO::PARAM_INT;
+					break;
+				case "double":
+					break;
+				case "string":
+					$type = PDO::PARAM_STR;
+					break;
+				case "array":
+					$value = implode(',', $value);
+					$type = PDO::PARAM_STR;
+					break;
+				case "object":
+					try
+					{
+						$value = (string)$value;
+					}
+					catch(Exception $e)
+					{
+						throw new QueryException('Could not convert object to string for query.', 0, $e);
+					}
+					$type = PDO::PARAM_STR;
+					break;
+				case "resource":
+				case "resource (closed)":
+					throw new QueryException('Cannot supply a resource to a query.');
+					break;
+				case "NULL":
+					$type = PDO::PARAM_NULL;
+					break;
+				default:
+					throw new QueryException('Unable to determine parameter type.');
+			}
+
+			//Check for a large value
+			if($type === PDO::PARAM_STR && strlen($value) > 4000)
+				$type = PDO::PARAM_LOB;
+
+			$statement->bindValue($name, $value, $type);
+		}
 	}
 	
 	
@@ -285,6 +413,7 @@ class Insert
 	 * @param string $column
 	 * @param string $value
 	 * @return $this
+	 * @throws QueryException
 	 */
 	public function addLiteralColumn($column, $value)
 	{
@@ -389,12 +518,30 @@ class Insert
 
 		return $this;
 	}
+
+	/**
+	 * @return bool
+	 */
+	public function isParameterized(): bool
+	{
+		return $this->parameterized;
+	}
+
+	/**
+	 * @param bool $parameterized
+	 * @return Insert
+	 */
+	public function setParameterized(bool $parameterized): Insert
+	{
+		$this->parameterized = $parameterized;
+		return $this;
+	}
 	
 	/**
-	 * @param Connection $connection
+	 * @param IConnection $connection
 	 * @return $this
 	 */
-	public function setConnection(Connection $connection)
+	public function setConnection(IConnection $connection)
 	{
 		$this->connection = $connection;
 		return $this;
@@ -419,7 +566,7 @@ class Insert
 		}
 		else
 		{
-			throw new QueryException('Data must be an instance of Staple_Query_DataSet, an instance of Staple_Query_Select or an array', Error::APPLICATION_ERROR);
+			throw new QueryException('Data must be an instance of \Staple\Query\DataSet, an instance of \Staple\Query\Select or an array', Error::APPLICATION_ERROR);
 		}
 		return $this;
 	}
