@@ -28,14 +28,16 @@ namespace Staple\Query;
 use DateTime;
 use Exception;
 use PDO;
+use PDOException;
 use PDOStatement;
 use Staple\Error;
+use Staple\Exception\ConfigurationException;
 use Staple\Exception\QueryException;
 use Staple\Pager;
 
 abstract class Query implements IQuery
 {
-	
+	const PARAM_NAME_VALID_CHARACTERS = '/[^a-zA-Z0-9_]/';
 	/**
 	 * Table to act upon.
 	 * @var mixed
@@ -52,12 +54,24 @@ abstract class Query implements IQuery
 	 * @var IConnection
 	 */
 	protected $connection;
+
+	/**
+	 * An array to hold all of the parameters of the query.
+	 * @var string[]
+	 */
+	protected $params = [];
 	
 	/**
 	 * An array of Where Clauses. The clauses are additive, using the AND  conjunction.
 	 * @var Condition[]
 	 */
 	protected $where = array();
+
+	/**
+	 * Set a flag if the query is to be parameterized.
+	 * @var bool
+	 */
+	protected $parameterized = true;
 
 	/**
 	 * @param string $table
@@ -191,26 +205,118 @@ abstract class Query implements IQuery
 	}
 
 	/**
+	 * Return the parameter list.
+	 * @return string[]
+	 */
+	public function getParams(): array
+	{
+		return $this->params;
+	}
+
+	/**
+	 * Set the value of a named parameter on the query.
+	 * @param string $paramName
+	 * @param mixed $value
+	 * @return $this
+	 * @throws QueryException
+	 */
+	public function setParam(string $paramName, $value): IQuery
+	{
+		if(is_resource($value)) {
+			throw new QueryException('Cannot use a resource as a value in query.');
+		}
+
+		if(substr($paramName,0,1) === ':')
+			$paramName = substr($paramName, 1);
+
+		$this->params[$paramName] = $value;
+		return $this;
+	}
+
+	/**
+	 * Check for name conflicts on param names
+	 * @param string $paramName
+	 * @return bool
+	 */
+	protected function checkParamNameConflict(string $paramName): bool
+	{
+		if(substr($paramName,0,1) === ':')
+			$paramName = substr($paramName, 1);
+
+		return array_key_exists($paramName, $this->params);
+	}
+
+	/**
+	 * Generate a unique parameter name
+	 * @param string $paramName
 	 * @return string
 	 */
-	abstract function build();
+	protected function generateIncrementalParamName(string $paramName): string
+	{
+		if(substr($paramName,0,1) === ':')
+			$paramName = substr($paramName, 1);
+
+		if($this->checkParamNameConflict($paramName))
+		{
+			$nameSections = explode('_', $paramName);
+			$finalSection = array_pop($nameSections);
+			if(ctype_digit($finalSection)) {
+				$finalSection = (int)$finalSection + 1;
+				array_push($nameSections, $finalSection);
+			}
+			else
+			{
+				array_push($nameSections, $finalSection, "1");
+			}
+
+			$newParamName = implode('_', $nameSections);
+			return $this->generateIncrementalParamName($newParamName);
+		}
+		else
+		{
+			return $paramName;
+		}
+	}
+
+	/**
+	 * @return bool
+	 */
+	public function isParameterized(): bool
+	{
+		return $this->parameterized;
+	}
+
+	/**
+	 * @param bool $parameterized
+	 * @return IQuery
+	 */
+	public function setParameterized(bool $parameterized): IQuery
+	{
+		$this->parameterized = $parameterized;
+		foreach($this->where as $condition) {
+			$condition->setParameterized($parameterized);
+		}
+		return $this;
+	}
+
+	/**
+	 * @param bool $parameterized
+	 * @return string
+	 */
+	abstract function build(bool $parameterized = null);
 	
 	/**
 	 * Executes the query and returns the result.
 	 * @param IConnection $connection - the database connection to execute the quote upon.
 	 * @return Statement | bool
-	 * @throws Exception
+	 * @throws QueryException
 	 */
 	public function execute(IConnection $connection = NULL)
 	{
 		if(isset($connection))
 			$this->setConnection($connection);
 
-		if($this->connection instanceof IConnection)
-		{
-			return $this->connection->query($this);
-		}
-		else
+		if(!($this->connection instanceof IConnection))
 		{
 			try 
 			{
@@ -220,12 +326,87 @@ abstract class Query implements IQuery
 			{
 				throw new QueryException('No Database Connection', Error::DB_ERROR);
 			}
-			if($this->connection instanceof IConnection)
-			{
-				return $this->connection->query($this);
-			}
 		}
-		return false;
+
+		if($this->isParameterized() !== true)
+		{
+			return $this->connection->query($this->build());
+		}
+		else
+		{
+			$statement = $this->connection->prepare($this->build(true));
+			if($statement !== false && $statement !== null)
+			{
+				$this->bindParametersToStatement($statement);
+				if($statement->execute() === true)
+					return $statement;
+				else
+					throw new QueryException(json_encode($statement->errorInfo()));
+			}
+			else
+				throw new QueryException('Failed to prepare statement for execution.');
+		}
+	}
+
+	/**
+	 * Bind the query parameters to the supplied statement object.
+	 * @param IStatement $statement
+	 * @throws QueryException
+	 */
+	private function bindParametersToStatement(&$statement)
+	{
+		foreach($this->getParams() as $name=>$value)
+		{
+			$varType = gettype($value);
+
+			//Don't bind nulls because they are converted to IS NULL by the query builder.
+			if($varType === 'NULL')
+			{
+				continue;
+			}
+
+			//Bind other types.
+			switch($varType)
+			{
+				case "boolean":
+					$type = PDO::PARAM_BOOL;
+					break;
+				case "integer":
+					$type = PDO::PARAM_INT;
+					break;
+				case "double":
+					break;
+				case "string":
+					$type = PDO::PARAM_STR;
+					break;
+				case "array":
+					$value = implode(',', $value);
+					$type = PDO::PARAM_STR;
+					break;
+				case "object":
+					try
+					{
+						$value = (string)$value;
+					}
+					catch(Exception $e) {
+						throw new QueryException('Could not convert object to string for query.', 0, $e);
+					}
+					$type = PDO::PARAM_STR;
+					break;
+				case "resource":
+				case "resource (closed)":
+					throw new QueryException('Cannot supply a resource to a query.');
+					break;
+				default:
+					throw new QueryException('Unable to determine parameter type.');
+			}
+
+			//Check for a large value
+			if($type === PDO::PARAM_STR && strlen($value) > 4000)
+				$type = PDO::PARAM_LOB;
+
+			$statement->bindValue($name, $value, $type);
+		}
 	}
 
 	/**
@@ -234,7 +415,7 @@ abstract class Query implements IQuery
 	 * @param IConnection $connection
 	 * @return Statement
 	 * @throws QueryException
-	 * @throws \PDOException
+	 * @throws PDOException
 	 */
 	public function prepareAndExecute(IConnection $connection = null)
 	{
@@ -284,6 +465,7 @@ abstract class Query implements IQuery
 	 * This method gets either the default framework connection or a predefined named connection.
 	 * @param string $namedConnection
 	 * @return Connection
+	 * @throws Exception
 	 */
 	public static function connection($namedConnection = NULL)
 	{
@@ -300,10 +482,23 @@ abstract class Query implements IQuery
 	}
 	
 	/*-----------------------------------------------WHERE CLAUSES-----------------------------------------------*/
-	
+
+	/**
+	 * @param Condition $where
+	 * @return $this
+	 * @throws QueryException
+	 */
 	public function addWhere(Condition $where)
 	{
 		$this->where[] = $where;
+
+		//Check for column name conflicts
+		if($where->getColumnJoin() === false && $where->isParameterized() === true && !is_null($where->getParamName()))
+		{
+			$newParamName = $this->generateIncrementalParamName($where->getParamName());
+			$this->setParam($newParamName, $where->getValue());
+			$where->setParamName($newParamName);
+		}
 		return $this;
 	}
 	
@@ -312,17 +507,82 @@ abstract class Query implements IQuery
 		$this->where = array();
 		return $this;
 	}
-	
-	public function whereCondition($column, $operator, $value, $columnJoin = NULL)
+
+	/**
+	 * Additive conditional WHERE clause
+	 * @param $column
+	 * @param $operator
+	 * @param $value
+	 * @param string $paramName
+	 * @param bool|null $columnJoin
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 */
+	public function where($column, $operator, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::get($column, $operator, $value, $columnJoin));
+		$this->addWhere(Condition::get($column, $operator, $value, $columnJoin, $paramName, Condition::SQL_AND, $parameterized));
+		return $this;
+	}
+
+	/**
+	 * Alternative (OR) conditional WHERE clause
+	 * @param $column
+	 * @param $operator
+	 * @param $value
+	 * @param bool|null $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 */
+	public function orWhere($column, $operator, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
+	{
+		$this->addWhere(Condition::get($column, $operator, $value, $columnJoin, $paramName, Condition::SQL_OR, $parameterized));
+		return $this;
+	}
+
+	/**
+	 * Additive Custom conditional WHERE (ex. less than, or greater than)
+	 * @param $column
+	 * @param $operator
+	 * @param $value
+	 * @param bool|null $columnJoin
+	 * @param string $paramName
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 * @deprecated
+	 */
+	public function whereCondition($column, $operator, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
+	{
+		return $this->where($column, $operator, $value, $columnJoin, $paramName, $parameterized);
+	}
+
+	/**
+	 * Alternative (OR) Custom conditional WHERE (ex. less than, or greater than)
+	 * @param $column
+	 * @param $operator
+	 * @param $value
+	 * @param bool|null $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 * @deprecated
+	 */
+	public function orWhereCondition($column, $operator, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
+	{
+		$this->orWhere($column, $operator, $value, $columnJoin, $paramName, $parameterized);
 		return $this;
 	}
 	
 	/**
-	 * An open ended where statement
+	 * An open ended raw where statement. Use this with caution as you are responsible for sanitizing the SQL that
+	 * is supplied to this method.
 	 * @param string | Select $statement
 	 * @return $this
+	 * @throws QueryException
 	 */
 	public function whereStatement($statement)
 	{
@@ -331,40 +591,86 @@ abstract class Query implements IQuery
 	}
 	
 	/**
-	 * SQL WHERE =
+	 * Additive SQL WHERE =
 	 * @param string $column
 	 * @param mixed $value
-	 * @param boolean $columnJoin
+	 * @param bool $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws QueryException
 	 */
-	public function whereEqual($column, $value, $columnJoin = NULL)
+	public function whereEqual($column, $value, bool $columnJoin = null, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::equal($column, $value, $columnJoin));
+		if(!isset($parameterized))
+			$parameterized = $this->isParameterized();
+		$this->addWhere(Condition::equal($column, $value, $columnJoin, $paramName, Condition::SQL_AND, $parameterized));
 		return $this;
 	}
 
 	/**
-	 * SQL WHERE <>
+	 * SQL OR WHERE =
 	 * @param string $column
 	 * @param mixed $value
-	 * @param boolean $columnJoin
+	 * @param bool $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws QueryException
 	 */
-	public function whereNotEqual($column, $value, $columnJoin = NULL)
+	public function orWhereEqual($column, $value, bool $columnJoin = null, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::notEqual($column, $value, $columnJoin));
+		if(!isset($parameterized))
+			$parameterized = $this->isParameterized();
+		$this->addWhere(Condition::equal($column, $value, $columnJoin, $paramName, Condition::SQL_OR, $parameterized));
+		return $this;
+	}
+
+	/**
+	 * Additive SQL WHERE <>
+	 * @param string $column
+	 * @param mixed $value
+	 * @param bool $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 */
+	public function whereNotEqual($column, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
+	{
+		$this->addWhere(Condition::notEqual($column, $value, $columnJoin, $paramName, Condition::SQL_AND, $parameterized));
+		return $this;
+	}
+
+	/**
+	 * Alternative SQL WHERE <>
+	 * @param string $column
+	 * @param mixed $value
+	 * @param bool|null $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
+	 * @return $this
+	 * @throws QueryException
+	 */
+	public function orWhereNotEqual($column, $value, bool $columnJoin = NULL, string $paramName = null, bool $parameterized = true)
+	{
+		$this->addWhere(Condition::notEqual($column, $value, $columnJoin, $paramName, Condition::SQL_OR, $parameterized));
 		return $this;
 	}
 	
 	/**
-	 * SQL LIKE Clause
+	 * Additive SQL LIKE Clause
 	 * @param string $column
 	 * @param mixed $value
+	 * @param bool $columnJoin
+	 * @param string $paramName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws QueryException
 	 */
-	public function whereLike($column, $value)
+	public function whereLike($column, $value, bool $columnJoin = null, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::like($column, $value));
+		$this->addWhere(Condition::like($column, $value, $columnJoin, $paramName, Condition::SQL_AND, $parameterized));
 		return $this;
 	}
 
@@ -372,11 +678,15 @@ abstract class Query implements IQuery
 	 * SQL NOT LIKE Clause
 	 * @param string $column
 	 * @param mixed $value
+	 * @param bool|null $columnJoin
+	 * @param string|null $paramName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws QueryException
 	 */
-	public function whereNotLike($column, $value)
+	public function whereNotLike($column, $value, bool $columnJoin = null, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::notLike($column, $value));
+		$this->addWhere(Condition::notLike($column, $value, $columnJoin, $paramName, Condition::SQL_AND, $parameterized));
 		return $this;
 	}
 	
@@ -384,6 +694,7 @@ abstract class Query implements IQuery
 	 * SQL IS NULL Clause
 	 * @param string $column
 	 * @return $this
+	 * @throws QueryException
 	 */
 	public function whereNull($column)
 	{
@@ -395,11 +706,14 @@ abstract class Query implements IQuery
 	 * SQL IN Clause
 	 * @param string $column
 	 * @param array | Select $values
+	 * @param string|null $paramName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws QueryException
 	 */
-	public function whereIn($column, $values)
+	public function whereIn($column, $values, string $paramName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::in($column, $values));
+		$this->addWhere(Condition::in($column, $values, $paramName, Condition::SQL_AND, $parameterized));
 		return $this;
 	}
 	
@@ -408,11 +722,15 @@ abstract class Query implements IQuery
 	 * @param string $column
 	 * @param mixed $start
 	 * @param mixed $end
+	 * @param string $startParamName
+	 * @param string $endParamName
+	 * @param bool $parameterized
 	 * @return $this
+	 * @throws Exception
 	 */
-	public function whereBetween($column, $start, $end)
+	public function whereBetween($column, $start, $end, string $startParamName = null, string $endParamName = null, bool $parameterized = true)
 	{
-		$this->addWhere(Condition::between($column, $start, $end));
+		$this->addWhere(Condition::between($column, $start, $end, $startParamName, $endParamName, Condition::SQL_AND, $parameterized));
 		return $this;
 	}
 	
@@ -484,6 +802,7 @@ abstract class Query implements IQuery
 	 *
 	 * @param string $table
 	 * @return static
+	 * @throws QueryException
 	 */
 	public static function table($table)
 	{
@@ -498,11 +817,13 @@ abstract class Query implements IQuery
 	 * @param IConnection $db
 	 * @param array | string $order
 	 * @param Pager | int $limit
+	 * @param bool $parameterized
 	 * @return Select
+	 * @throws QueryException
 	 */
-	public static function select($table = NULL, array $columns = NULL, IConnection $db = NULL, $order = NULL, $limit = NULL)
+	public static function select($table = NULL, array $columns = NULL, IConnection $db = NULL, $order = NULL, $limit = NULL, bool $parameterized = null)
 	{
-		return new Select($table, $columns, $db, $order, $limit);
+		return new Select($table, $columns, $db, $order, $limit, $parameterized);
 	}
 
 	/**
@@ -512,11 +833,13 @@ abstract class Query implements IQuery
 	 * @param array $data
 	 * @param IConnection
 	 * @param string $priority
+	 * @param bool $parameterized
 	 * @return Insert
+	 * @throws QueryException
 	 */
-	public static function insert($table = NULL, $data = NULL, IConnection $db = NULL, $priority = NULL)
+	public static function insert($table = NULL, $data = NULL, IConnection $db = NULL, $priority = NULL, bool $parameterized = null)
 	{
-		return new Insert($table, $data, $db, $priority);
+		return new Insert($table, $data, $db, $priority, $parameterized);
 	}
 
 	/**
@@ -527,11 +850,13 @@ abstract class Query implements IQuery
 	 * @param IConnection $db
 	 * @param array | string $order
 	 * @param Pager | int $limit
+	 * @param bool $parameterized
 	 * @return Update
+	 * @throws QueryException
 	 */
-	public static function update($table = NULL, array $data = NULL, IConnection $db = NULL, $order = NULL, $limit = NULL)
+	public static function update($table = NULL, array $data = NULL, IConnection $db = NULL, $order = NULL, $limit = NULL, bool $parameterized = null)
 	{
-		return new Update($table, $data, $db, $order, $limit);
+		return new Update($table, $data, $db, $order, $limit, $parameterized);
 	}
 
 	/**
@@ -539,11 +864,13 @@ abstract class Query implements IQuery
 	 *
 	 * @param string $table
 	 * @param IConnection $db
+	 * @param bool $parameterized
 	 * @return Delete
+	 * @throws QueryException
 	 */
-	public static function delete($table = NULL, IConnection $db = NULL)
+	public static function delete($table = NULL, IConnection $db = NULL, bool $parameterized = null)
 	{
-		return new Delete($table, $db);
+		return new Delete($table, $db, $parameterized);
 	}
 
 	/**
@@ -552,6 +879,7 @@ abstract class Query implements IQuery
 	 * @param array $queries
 	 * @param IConnection $db
 	 * @return Union
+	 * @throws QueryException
 	 */
 	public static function union(array $queries = array(), IConnection $db = NULL)
 	{
@@ -575,6 +903,7 @@ abstract class Query implements IQuery
      * @param string | Query $statement
      * @param IConnection $connection
      * @return Statement
+	 * @throws ConfigurationException
      */
 	public static function raw($statement, IConnection $connection = NULL)
 	{
@@ -592,6 +921,7 @@ abstract class Query implements IQuery
 	 * @param array $driverOptions
 	 * @return PDOStatement
 	 * @throws QueryException
+	 * @throws ConfigurationException
 	 */
 	public static function procedure($procedureName, array $parameters = NULL, IConnection $connection = NULL, array $driverOptions = [])
 	{
@@ -740,5 +1070,17 @@ abstract class Query implements IQuery
 		ksort($parameters);
 
 		return 'EXEC '.$procedureName.$params;
+	}
+
+	/**
+	 * Eliminate any invalid characters from the parameter name.
+	 * @param string $paramName
+	 * @return string
+	 */
+	public static function sanitizeParamName(string $paramName): string
+	{
+		if(substr($paramName,0,1) === ':')
+			$paramName = substr($paramName, 1);
+		return preg_replace(self::PARAM_NAME_VALID_CHARACTERS, '_', $paramName);
 	}
 }
