@@ -25,6 +25,7 @@ namespace Staple\Query;
 
 use PDO;
 use PDOStatement;
+use Staple\Exception\ConfigurationException;
 
 class Statement extends PDOStatement implements IStatement
 {
@@ -34,9 +35,18 @@ class Statement extends PDOStatement implements IStatement
      */
     protected $driver;
 
-	/*
+    /**
+	 * The data store connection.
+     * @var Connection
+     */
+    protected $connection;
+
+	/**
 	 * Magic method to fake MySQLi property functions
 	 * @deprecated
+	 * @param string $name
+	 * @return int|null
+	 * @throws ConfigurationException
 	 */
 	public function __get($name)
 	{
@@ -69,6 +79,24 @@ class Statement extends PDOStatement implements IStatement
     }
 
 	/**
+	 * @return IConnection
+	 */
+	public function getConnection(): IConnection
+	{
+		return $this->connection;
+	}
+
+	/**
+	 * @param IConnection $connection
+	 * @return IStatement
+	 */
+	public function setConnection(IConnection $connection): IStatement
+	{
+		$this->connection = $connection;
+		return $this;
+	}
+
+	/**
 	 * Mysqli style associative array fetch style
 	 * @return mixed
 	 * @deprecated
@@ -91,6 +119,7 @@ class Statement extends PDOStatement implements IStatement
     /**
      * Returns the number of rows found in the previous query.
      * @return string
+	 * @throws ConfigurationException
      */
     public function foundRows()
     {
@@ -108,6 +137,7 @@ class Statement extends PDOStatement implements IStatement
 	/**
 	 * Override the PDOStatement rowCount() method to return
 	 * @return int
+	 * @throws ConfigurationException
 	 */
 	public function rowCount()
 	{
@@ -119,4 +149,16 @@ class Statement extends PDOStatement implements IStatement
 				return parent::rowCount();
 		}
 	}
+
+	/**
+	 * @param array|null $input_parameters
+	 * @return bool
+	 */
+	public function execute($input_parameters = null)
+	{
+		$this->getConnection()->addQueryToLog($this->queryString);
+		return parent::execute($input_parameters);
+	}
+
+
 }
