@@ -72,10 +72,10 @@ class Union
 	/**
 	 * Constructor accepts an array of Select elements and a database connection.
 	 * @param array $queries
-	 * @param Connection $connection
+	 * @param IConnection $connection
 	 * @throws QueryException
 	 */
-	public function __construct(array $queries = array(), Connection $connection = NULL)
+	public function __construct(array $queries = array(), IConnection $connection = NULL)
 	{
 		//Process Database connection
 		if($connection instanceof Connection)
@@ -190,6 +190,56 @@ class Union
 	{
 		$this->connection = $connection;
 		return $this;
+	}
+
+	/**
+	 * @return array
+	 */
+	public function getParams()
+	{
+		$params = [];
+		foreach($this->getQueries() as $query)
+		{
+			$queryParams = $query->getParams();
+			foreach($queryParams as $name=>$value)
+			{
+				$params[$this->generateIncrementalParamName($name, $params)] = $value;
+			}
+		}
+		return $params;
+	}
+
+	/**
+	 * Generate a unique parameter name
+	 * @param string $paramName
+	 * @param array $masterParamList
+	 * @return string
+	 */
+	protected function generateIncrementalParamName(string $paramName, array $masterParamList): string
+	{
+		if(substr($paramName,0,1) === ':')
+			$paramName = substr($paramName, 1);
+
+		if(array_key_exists($paramName, $masterParamList))
+		{
+			$nameSections = explode('_', $paramName);
+			$finalSection = array_pop($nameSections);
+			if(ctype_digit($finalSection)) {
+				$finalSection = (int)$finalSection + 1;
+				array_push($nameSections, $finalSection);
+			}
+			else
+			{
+				array_push($nameSections, $finalSection, "1");
+			}
+
+			$newParamName = implode('_', $nameSections);
+			return $this->generateIncrementalParamName($newParamName, $masterParamList);
+		}
+		else
+		{
+			return $paramName;
+		}
 	}
 
 	/**
@@ -485,7 +535,7 @@ class Union
 		{
 			try 
 			{
-				$this->db = Connection::get();
+				$this->connection = Connection::get();
 			}
 			catch (Exception $e)
 			{
