@@ -24,22 +24,29 @@
 namespace Staple\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Staple\Exception\QueryException;
 use Staple\Query\Insert;
 use Staple\Query\Select;
 use Staple\Query\MockConnection;
+use DateTime;
 
 class InsertTest extends TestCase
 {
-	const INSERT_IGNORE_STATEMENT = "INSERT HIGH_PRIORITY IGNORE \nINTO schemestaple.categories (id,name,summary,rank,cat,created) \nVALUES (1,'Test','This is a test and only a test.','3.14',2,'2010-01-01 00:00:00') ";
-	const INSERT_SELECT_STATEMENT = "INSERT \nINTO schemestaple.categories \nSELECT\nname, summary, rank, category_id AS cat, created \nFROM schemestaple.article_categories\nWHERE id = articles.cat";
-	const INSERT_SELECT_COLUMNS_STATEMENT = "INSERT \nINTO schemestaple.categories (name,summary,rank,cat,created) \nSELECT\nname, summary, rank, category_id AS cat, created \nFROM schemestaple.article_categories\nWHERE id = articles.cat";
+	const INSERT_IGNORE_STATEMENT = "INSERT HIGH_PRIORITY OR IGNORE \nINTO categories (id, name, summary, rank, cat, created) \nVALUES (:id, :name, :summary, :rank, :cat, :created) ";
+	const INSERT_SELECT_STATEMENT = "INSERT \nINTO categories \nSELECT\nname, summary, rank, category_id AS cat, created \nFROM article_categories\nWHERE id = articles.cat";
+	const INSERT_SELECT_COLUMNS_STATEMENT = "INSERT \nINTO categories (name, summary, rank, cat, created) \nSELECT\nname, summary, rank, category_id AS cat, created \nFROM article_categories\nWHERE id = articles.cat";
 	const INSERT_SCHEMA_STATEMENT = "INSERT \nINTO myschema.categories \nSELECT\nname, summary, rank, category_id AS cat, created \nFROM myschema.article_categories\nWHERE id = articles.cat";
+	const INSERT_DATA_STATEMENT = "INSERT \nINTO customer (first_name, last_name, address, city, state, zip, active, created, modified) \nVALUES (:first_name, :last_name, :address, :city, :state, :zip, :active, :created, NOW()) ";
 
 	private function getMockConnection()
 	{
 		return new MockConnection('sqlite::memory:');
 	}
 
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
 	public function testInsertSelect()
 	{
 		//Setup the database
@@ -63,6 +70,10 @@ class InsertTest extends TestCase
 		$this->assertEquals(self::INSERT_SELECT_STATEMENT, (string)$insertSelect);
 	}
 
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
 	public function testInsertSelectWithCustomColumnList()
 	{
 		//Setup the database
@@ -93,24 +104,37 @@ class InsertTest extends TestCase
 		$this->assertEquals(self::INSERT_SELECT_COLUMNS_STATEMENT, (string)$insertSelect);
 	}
 
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
 	public function testInsertIgnore()
 	{
+		$data = ['id'	=>	1,
+			'name'		=>	'Test',
+			'summary'	=>	'This is a test and only a test.',
+			'rank'		=>	3.14,
+			'cat'		=>	2,
+			'created' 	=>	new DateTime('2010-01-01 00:00:00')
+		];
+
 		//Create the Query
 		$insert = Insert::create()
 			->setTable('categories')
 			->setPriority(Insert::HIGH)
 			->setIgnore(true)
-			->addData(['id'	=>	1,
-					   'name'		=>	'Test',
-					   'summary'	=>	'This is a test and only a test.',
-					   'rank'		=>	3.14,
-					   'cat'		=>	2,
-					   'created' 	=>	new \DateTime('2010-01-01 00:00:00')
-			]);
+			->addData($data);
 
 		$this->assertEquals(self::INSERT_IGNORE_STATEMENT,(string)$insert);
+
+		//Check the param array
+		$this->assertEquals($data, $insert->getParams());
 	}
 
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
 	public function testInsertWithSchema()
 	{
 		//Setup the database
@@ -134,5 +158,36 @@ class InsertTest extends TestCase
 			->setData($select);
 
 		$this->assertEquals(self::INSERT_SCHEMA_STATEMENT, (string)$insertSelect);
+	}
+
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
+	public function testInsertWithUserData()
+	{
+		//Setup the database
+		$this->getMockConnection();
+
+		$data = [
+			'first_name' => 'Jim',
+			'last_name' => 'Bob',
+			'address' => '123 Test Street',
+			'city' => 'Brightville',
+			'state' => 'Michigan',
+			'zip' => '12345',
+			'active' => true,
+			'created' => time(),
+		];
+
+		$insert = Insert::create()
+			->setTable('customer')
+			->setData($data)
+			->addLiteralColumn('modified', 'NOW()');
+
+		$this->assertEquals(self::INSERT_DATA_STATEMENT, (string)$insert);
+
+		//Check the param array
+		$this->assertEquals($data, $insert->getParams());
 	}
 }
