@@ -24,12 +24,16 @@
 namespace Staple\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Staple\Exception\ConfigurationException;
+use Staple\Exception\QueryException;
 use Staple\Query\Connection;
 use Staple\Query\Query;
 use Staple\Query\MockConnection;
 
 class QueryTest extends TestCase
 {
+	const PARAMETERIZED_QUERY_STRING = "SELECT\n* \nFROM customers\nWHERE city = :city AND age = :age";
+
 	private function getMockConnection()
 	{
 		return new MockConnection(NULL);
@@ -39,6 +43,10 @@ class QueryTest extends TestCase
 		return new MockConnection('sqlite::memory:');
 	}
 
+	/**
+	 * @test
+	 * @throws \Staple\Exception\QueryException
+	 */
 	public function testMySQLQueryCreation()
 	{
 		//Setup
@@ -54,9 +62,9 @@ class QueryTest extends TestCase
 		$this->assertInstanceOf('\Staple\Query\Select',$select);
 		$this->assertEquals("SELECT\n* \nFROM customers",(string)$select);
 		$this->assertInstanceOf('\Staple\Query\Insert',$insert);
-		$this->assertEquals("INSERT \nINTO customers (id) \nVALUES (1) ",(string)$insert);
+		$this->assertEquals("INSERT \nINTO customers (id) \nVALUES (:id) ",(string)$insert);
 		$this->assertInstanceOf('\Staple\Query\Update',$update);
-		$this->assertEquals("UPDATE customers\nSET name='Larry'",(string)$update);
+		$this->assertEquals("UPDATE customers\nSET name=:name",(string)$update);
 		$this->assertInstanceOf('\Staple\Query\Delete',$delete);
 		$this->assertEquals("DELETE FROM customers",(string)$delete);
 	}
@@ -74,6 +82,8 @@ class QueryTest extends TestCase
 	/**
 	 * Test that we can construct a stored procedure call to MySQL
 	 * @test
+	 * @throws QueryException
+	 * @throws ConfigurationException
 	 */
 	public function testStoredMySqlProcedureConstruction()
 	{
@@ -110,6 +120,8 @@ class QueryTest extends TestCase
 	/**
 	 * Test that we can construct a stored procedure call to SQL Server
 	 * @test
+	 * @throws QueryException
+	 * @throws ConfigurationException
 	 */
 	public function testStoredSqlSrvProcedureConstruction()
 	{
@@ -119,5 +131,43 @@ class QueryTest extends TestCase
 		Query::procedure('GetCustomers',['first_name'=>'John','last_name'=>'Smith'],$connection);
 
 		$this->assertEquals('EXEC GetCustomers @first_name = ?, @last_name = ?', $connection->getLastQuery());
+	}
+
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
+	public function testParameterizedQuery()
+	{
+		$connection = $this->getMockConnection();
+		$connection->setDriver(Connection::DRIVER_MYSQL);
+
+		$query = Query::select('customers', null, $connection)
+			->whereEqual('city', 'Memphis')
+			->whereEqual('age', 20);
+
+		$queryString = $query->build();
+
+		$this->assertEquals(self::PARAMETERIZED_QUERY_STRING, $queryString);
+
+		$params = $query->getParams();
+
+		$this->assertEquals(['city'=>'Memphis','age'=>20], $params);
+	}
+
+	/**
+	 * @test
+	 * @throws QueryException
+	 */
+	public function testParameterizedQueryToStringConversion()
+	{
+		$connection = $this->getMockConnection();
+		$connection->setDriver(Connection::DRIVER_MYSQL);
+
+		$query = Query::select('customers', null, $connection)
+			->whereEqual('city', 'Memphis')
+			->whereEqual('age', 20);
+
+		$this->assertEquals(self::PARAMETERIZED_QUERY_STRING, (string)$query);
 	}
 }
