@@ -66,7 +66,16 @@ return [
 		'username'	=>	NULL,
 		'password'	=>	NULL,
 		'db'		=>	'staple',
-		'schema'	=>	'schemestaple',
+		//'schema'	=>	'schemestaple',
+		'options'	=>	[],     		//Specify any additional options here
+	],
+
+	'db-mysql'	=>	[
+		'driver'	=>	'mysql',    	//specify the database connection type
+		'host'		=> 	'127.0.0.1',
+		'username'	=>	'root',
+		'password'	=>	NULL,
+		'db'		=>	'staple',
 		'options'	=>	[],     		//Specify any additional options here
 	],
 
